Make Gluon backend rep common
